Create HibernateDao for three tables and create the JunitTest for each Dao class

+++ src/main/java/org/antontech/model/Product.java
@@ -8,22 +8,22 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
     private String description;
     @Column(name="user_id")
-    private int user_id;
+    private long user_id;
 
     public Product() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -43,7 +43,7 @@ public class Product {
         this.description = description;
     }
 
-    public int getUser_id() {return user_id;}
+    public long getUser_id() {return user_id;}
 
-    public void setUser_id(int user_id) {this.user_id = user_id;}
+    public void setUser_id(long user_id) {this.user_id = user_id;}
 }
